+++ src/components/Thumbnail.js
@@ -1,12 +1,18 @@
+import React from "react";
 import styled from "styled-components";
 
-const Thumbnail = styled.div`
+const Container = styled.div`
   background: url(${props => props.image}) no-repeat center;
+  border: #aaa solid 1px;
   background-size: cover;
-  border-radius: 20px;
+  border-radius: ${props => props.size}px;
   display: inline-block;
-  height: 25px;
-  width: 25px;
+  height: ${props => props.size}px;
+  width: ${props => props.size}px;
 `;
+
+const Thumbnail = ({ size = 25, ...props }) => (
+  <Container size={size} {...props} />
+);
 
 export default Thumbnail;