+++ test/Icon.test.js
@@ -1,14 +1,10 @@
 import React from "react";
-import { shallow, mount } from "enzyme";
+import { shallow } from "enzyme";
 
-import Icon, { AsyncPath } from "../src/components/Icon";
+import Icon from "../src/components/Icon";
 
 test("renders without crashing", async () => {
-  const component = mount(<AsyncPath icon="checkmark" />);
-  expect(component.find("path")).toHaveLength(0);
-
-  await component.instance().componentDidMount();
-  component.update();
+  const component = shallow(<Icon icon="checkmark" />);
 
   expect(component.find("path")).toHaveLength(1);
 });