+++ CHANGELOG.md
@@ -5,8 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.7] - 2018-06-22
 ### Added
 - The `Button` component.
+- The optional `size` prop to the `Thumbnail` component.
+
+### Changed
+- Removed the dynamic import of the `icons.json` file since it's so much smaller now.
 
 ## [0.0.6] - 2018-06-19
 ### Changed